// src/types/base.type.ts
export interface BreakpointSizesType {
  xs?: number
  xs2?: number
  sm?: number
  sm2?: number
  md?: number
  md2?: number
  lg?: number
  lg2?: number
  xl?: number
  xl2?: number
  xl3?: number
  xl4?: number
  xl5?: number
}

export type ColSpanType =
  | '1'
  | '2'
  | '3'
  | '4'
  | '5'
  | '6'
  | '7'
  | '8'
  | '9'
  | '10'
  | '11'
  | '12'

export type GridDirectionType =
  | 'row'
  | 'row-reverse'
  | 'column'
  | 'column-reverse'
export type JustifyContentType =
  | 'flex-start'
  | 'center'
  | 'flex-end'
  | 'space-between'
  | 'space-around'
  | 'space-evenly'
export type AlignItemsType =
  | 'flex-start'
  | 'center'
  | 'flex-end'
  | 'stretch'
  | 'baseline'

export type Sizes = 'small' | 'medium' | 'large'

export type HtmlTags = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'span' | 'p'